Move files into subfolder for easier importlib loading
