+++ utils/memoryStream.py
